refactor(test): group credential coverage
